+++ src/main/java/kitchenpos/menu/response/MenuProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.response;
+package kitchenpos.menu.response;
 
 import kitchenpos.menu.domain.MenuProduct;
