+++ src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -2,8 +2,8 @@ package kitchenpos.menu.ui;
 
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.request.MenuCreateRequest;
-import kitchenpos.response.MenuResponse;
+import kitchenpos.menu.request.MenuCreateRequest;
+import kitchenpos.menu.response.MenuResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
