+++ src/main/java/kitchenpos/menugroup/ui/MenuGroupRestController.java
@@ -1,9 +1,9 @@
 package kitchenpos.menugroup.ui;
 
+import kitchenpos.menu.request.MenuGroupCreateRequest;
+import kitchenpos.menu.response.MenuGroupResponse;
 import kitchenpos.menugroup.application.MenuGroupService;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.request.MenuGroupCreateRequest;
-import kitchenpos.response.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
