+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
+import kitchenpos.menu.request.MenuGroupCreateRequest;
 import kitchenpos.menugroup.application.MenuGroupService;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.request.MenuGroupCreateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
