+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -5,10 +5,10 @@ import kitchenpos.dao.ProductDao;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.request.MenuCreateRequest;
+import kitchenpos.menu.request.MenuProductDto;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
-import kitchenpos.request.MenuCreateRequest;
-import kitchenpos.request.MenuProductDto;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
