+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -7,6 +7,8 @@ import kitchenpos.dao.ProductDao;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.request.MenuCreateRequest;
+import kitchenpos.menu.request.MenuProductDto;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
@@ -14,8 +16,6 @@ import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.application.TableService;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
-import kitchenpos.request.MenuCreateRequest;
-import kitchenpos.request.MenuProductDto;
 import kitchenpos.request.OrderTableCreateRequest;
 import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.tablegroup.application.TableGroupService;
